storage: extract core-boot install steps into CoreBootInstaller
The snapd "core boot" (TPM/FDE) install steps lived on StorageController
but are install-phase calls into snapd's v2/systems/<label> API, not
storage-layout concerns -- they only shared its bridge state. Move them
into a new CoreBootInstaller in subiquity/server/snapd/, passing that
state via a CoreBootInstallPlan dataclass the controller builds.

The main benefit is a smaller storage controller. The snapd-system install bits
now consistenly reside InstallController.
