+++ subiquity/server/controllers/tests/test_storage.py
@@ -88,6 +88,7 @@ from subiquity.server.controllers.storage import (
 from subiquity.server.dryrun import DRConfig
 from subiquity.server.snapd import api as snapdapi
 from subiquity.server.snapd import types as snapdtypes
+from subiquity.server.snapd.core_boot_installer import CoreBootInstaller
 from subiquity.server.snapd.info import SnapdInfo
 from subiquity.server.snapd.system_getter import SystemGetter
 from subiquity.server.snapd.types import VolumesAuth, VolumesAuthMode
@@ -581,42 +582,6 @@ class TestSubiquityControllerStorage(IsolatedAsyncioTestCase):
             await self.ctrler._probe_once(restricted=True)
         self.assertIsNone(self.ctrler.queued_probe_data)
         load.assert_not_called()
-
-    async def test_setup_encryption__passes_keyboard_config(self):
-        self.ctrler._info = mock.Mock()
-        self.ctrler._info.label = "prefer-encrypted"
-        self.ctrler._info.system.volumes = {"vol-key": mock.Mock()}
-        self.ctrler._on_volume = mock.Mock()
-        self.ctrler._volumes_auth = None
-
-        kb = self.app.base_model.keyboard
-        kb.setting.layout = "fr"
-        kb.setting.variant = "azerty"
-        kb.setting.toggle = "alt_shift_toggle"
-
-        with mock.patch.object(
-            snapdapi,
-            "post_and_wait",
-            return_value=mock.Mock(encrypted_devices={}),
-        ) as m_post:
-            await self.ctrler.setup_encryption()
-
-        # post_and_wait is called as:
-        #   post_and_wait(client, meth, request, ann=...)
-        # so:
-        #   args[0] is the snapdapi client
-        #   args[1] is the systems[label].POST bound method
-        #   args[2] is the SystemActionRequest
-        req = m_post.call_args.args[2]
-        self.assertEqual(
-            snapdtypes.KeyboardConfig(
-                model="pc105",
-                layout="fr",
-                variant="azerty",
-                options=["grp:alt_shift_toggle"],
-            ),
-            req.keyboard_config,
-        )
 
     fw_lenovo = {
         "bios-vendor": "LENOVO",
@@ -1744,115 +1709,6 @@ class TestSubiquityControllerStorage(IsolatedAsyncioTestCase):
 
         self.assertEqual(leading_gap.offset, gap.offset)
         self.assertEqual(part.size + leading_gap.size + trailing_gap.size, gap.size)
-
-    async def test_fetch_core_boot_recovery_key(self):
-        self.app.snapd = AsyncSnapd(get_fake_connection())
-        self.app.snapdapi = snapdapi.make_api_client(self.app.snapd)
-        self.ctrler._info = mock.Mock(label="my-label")
-
-        with mock.patch.object(
-            self.ctrler.model, "set_core_boot_recovery_key"
-        ) as m_set_key:
-            await self.ctrler.fetch_core_boot_recovery_key()
-
-        m_set_key.assert_called_once_with("my-recovery-key")
-
-    @mock.patch("subiquity.server.mounter.Mounter.bind_mounted")
-    @mock.patch.object(Path, "mkdir", mock.Mock())
-    async def test_snapd_target_preseed(self, m_bind_mounted):
-        self.ctrler._info = mock.Mock(label="mock-label")
-
-        with mock.patch.object(snapdapi, "post_and_wait") as mock_post:
-            await self.ctrler.snapd_target_preseed(Path("/target"))
-
-        expected_mounted_calls = [
-            mock.call(Path("/dev"), Path("/target/dev")),
-            mock.call(Path("/proc"), Path("/target/proc")),
-            mock.call(Path("/sys"), Path("/target/sys")),
-            mock.call(
-                Path("/sys/kernel/security"), Path("/target/sys/kernel/security")
-            ),
-            mock.call(Path("/var/lib/snapd/seed"), Path("/target/var/lib/snapd/seed")),
-        ]
-
-        self.assertEqual(expected_mounted_calls, m_bind_mounted.call_args_list)
-
-        mock_post.assert_called_once()
-
-    async def test_finish_install(self):
-        self.app.snapdapi = snapdapi.make_api_client(AsyncSnapd(get_fake_connection()))
-        variation_info = VariationInfo(
-            name="mock",
-            label="mock-label",
-            system=snapdtypes.SystemDetails(
-                label="mock-label",
-                volumes={
-                    "mockVol": snapdtypes.Volume(
-                        schema="mock", structure=None, bootloader="grub"
-                    ),
-                },
-                model=snapdtypes.Model(
-                    architecture="mock-arch",
-                    snaps=[
-                        snapdtypes.ModelSnap(
-                            name="MockKernel",
-                            type=snapdtypes.ModelSnapType.KERNEL,
-                            presence=snapdtypes.PresenceValue.REQUIRED,
-                            components={
-                                "nvidia-510-uda-ko": snapdtypes.PresenceValue.OPTIONAL,
-                                "nvidia-510-uda-user": snapdtypes.PresenceValue.OPTIONAL,
-                                "foo": snapdtypes.PresenceValue.OPTIONAL,
-                                "bar": snapdtypes.PresenceValue.OPTIONAL,
-                            },
-                            default_channel="foo",
-                            id="bar",
-                        ),
-                        snapdtypes.ModelSnap(
-                            name="MockApp1",
-                            type=snapdtypes.ModelSnapType.APP,
-                            presence=snapdtypes.PresenceValue.REQUIRED,
-                            default_channel="foo",
-                            id="bar",
-                        ),
-                        snapdtypes.ModelSnap(
-                            name="MockApp2",
-                            type=snapdtypes.ModelSnapType.APP,
-                            presence=snapdtypes.PresenceValue.OPTIONAL,
-                            default_channel="foo",
-                            id="bar",
-                        ),
-                    ],
-                ),
-                available_optional=snapdtypes.AvailableOptional(
-                    snaps=["MockApp2"],
-                    components={
-                        "MockKernel": [
-                            "nvidia-510-uda-ko",
-                            "nvidia-510-uda-user",
-                            "foo",
-                            "bar",
-                        ]
-                    },
-                ),
-            ),
-        )
-        self.ctrler._info = variation_info
-        with mock.patch.object(snapdapi, "post_and_wait") as mock_post:
-            await self.ctrler.finish_install(
-                context=self.ctrler.context,
-                kernel_components=["nvidia-510-uda-ko", "nvidia-510-uda-user"],
-            )
-        mock_post.assert_called_once()
-
-        # Assert installing all optional snaps but only the requested components
-        expected_optional_install = snapdtypes.OptionalInstall(
-            all=False,
-            components={"MockKernel": ["nvidia-510-uda-ko", "nvidia-510-uda-user"]},
-            snaps=variation_info.system.available_optional.snaps,
-        )
-        actual = mock_post.call_args.args[2].optional_install
-
-        self.assertEqual(expected_optional_install, actual)
 
 
 class TestRunAutoinstallGuided(IsolatedAsyncioTestCase):
@@ -3329,9 +3185,9 @@ class TestCoreBootInstallMethods(IsolatedAsyncioTestCase):
         )
         await self.ctrler.guided_core_boot(disk, choice)
         if va_expected is None:
-            self.assertIsNone(self.ctrler._volumes_auth)
+            self.assertIsNone(self.ctrler.core_boot_data.volumes_auth)
         else:
-            self.assertEqual(va_expected, self.ctrler._volumes_auth)
+            self.assertEqual(va_expected, self.ctrler.core_boot_data.volumes_auth)
         [part1, part2] = disk.partitions()
         self.assertEqual(part1.offset, 1 << 20)
         self.assertEqual(part1.size, 1 << 30)
@@ -3373,7 +3229,7 @@ class TestCoreBootInstallMethods(IsolatedAsyncioTestCase):
             capability=self.capability,
         )
         await self.ctrler.guided_core_boot(disk, choice)
-        self.assertIsNone(self.ctrler._volumes_auth)
+        self.assertIsNone(self.ctrler.core_boot_data.volumes_auth)
         [part] = disk.partitions()
         self.assertEqual(reused_part, part)
         self.assertEqual(reused_part.wipe, "superblock")
@@ -3401,7 +3257,7 @@ class TestCoreBootInstallMethods(IsolatedAsyncioTestCase):
             capability=self.capability,
         )
         await self.ctrler.guided_core_boot(disk, choice)
-        self.assertIsNone(self.ctrler._volumes_auth)
+        self.assertIsNone(self.ctrler.core_boot_data.volumes_auth)
         [part] = disk.partitions()
         self.assertEqual(existing_part, part)
         self.assertEqual(existing_part.wipe, None)
@@ -3435,7 +3291,7 @@ class TestCoreBootInstallMethods(IsolatedAsyncioTestCase):
             capability=self.capability,
         )
         await self.ctrler.guided_core_boot(disk, choice)
-        self.assertIsNone(self.ctrler._volumes_auth)
+        self.assertIsNone(self.ctrler.core_boot_data.volumes_auth)
         [bios_part, part] = disk.partitions()
         self.assertEqual(part.offset, 2 << 20)
         self.assertEqual(part.partition_name, "ptname")
@@ -3492,6 +3348,8 @@ class TestCoreBootInstallMethods(IsolatedAsyncioTestCase):
         device_map = {p.id: random_string() for p in disk.partitions()}
         self.ctrler.update_devices(device_map)
 
+        installer = CoreBootInstaller(self.app, self.ctrler.core_boot_data)
+
         with mock.patch.object(
             snapdapi, "post_and_wait", new_callable=mock.AsyncMock
         ) as mocked:
@@ -3500,7 +3358,10 @@ class TestCoreBootInstallMethods(IsolatedAsyncioTestCase):
                     snapdtypes.Role.SYSTEM_DATA: "enc-system-data",
                 },
             )
-            await self.ctrler.setup_encryption(context=self.ctrler.context)
+            await installer.setup_encryption(
+                context=self.ctrler.context,
+                apply_encrypted_devices=self.ctrler.apply_encrypted_devices,
+            )
 
         # setup_encryption mutates the filesystem model objects to
         # reference the newly created encrypted objects so re-read the
@@ -3511,7 +3372,7 @@ class TestCoreBootInstallMethods(IsolatedAsyncioTestCase):
         with mock.patch.object(
             snapdapi, "post_and_wait", new_callable=mock.AsyncMock
         ) as mocked:
-            await self.ctrler.finish_install(
+            await installer.finish_install(
                 context=self.ctrler.context, kernel_components=[]
             )
         mocked.assert_called_once()
@@ -3813,9 +3674,8 @@ class TestCalculateEntropy(IsolatedAsyncioTestCase):
         label = self.ctrler._info.label
         self.app.snapd = AsyncSnapd(get_fake_connection())
 
-        with mock.patch(
-            "subiquity.server.controllers.storage.snapdapi.make_api_client",
-            return_value=self.app.snapdapi,
+        with mock.patch.object(
+            snapdapi, "make_api_client", return_value=self.app.snapdapi
         ):
             with mock.patch.object(
                 self.app.snapdapi.v2.systems[label],
@@ -3848,9 +3708,8 @@ class TestCalculateEntropy(IsolatedAsyncioTestCase):
         label = self.ctrler._info.label
         self.app.snapd = AsyncSnapd(get_fake_connection())
 
-        with mock.patch(
-            "subiquity.server.controllers.storage.snapdapi.make_api_client",
-            return_value=self.app.snapdapi,
+        with mock.patch.object(
+            snapdapi, "make_api_client", return_value=self.app.snapdapi
         ):
             with mock.patch.object(
                 self.app.snapdapi.v2.systems[label],
